Oct 23, 2022, 2:47 PM

+++ Core Contracts/TreasuryV1.sol
@@ -46,7 +46,7 @@ contract HarmoniaDAOTreasury{
     }
 
     //Public callable functions
-    function ReceiveRegisteredAsset(uint AssetId, uint amount) external {
+    function ReceiveRegisteredAsset(uint8 AssetId, uint amount) external {
         ERC20(RegisteredAssets[AssetId].TokenAddress).transferFrom(msg.sender, address(this), amount);
         uint256 NewBalance = UpdateERC20Balance(AssetId);
